refactor: 챗봇 API 응답 처리 로직 개선
- useSendMessageMutation 훅에 성공/실패 처리 로직 통합
- ChatBotWindow에서 onSuccess/onError 콜백 제거로 코드 간소화
- 비즈니스 로직과 UI 로직의 명확한 분리
